pyproject.toml
[tool.poetry]
description = ""
name = "deployer"
version = "0.1.0"
authors = []
packages = [{include = "deployer"}]

[tool.poetry.dependencies]
python = "~3.10"

djangorestframework = "3.13.1"

[tool.poetry.dev-dependencies]
bandit = "^1.7.0"
black = "^22.3"
isort = "^5.11.4"
pylama = "^7.7.1"
pylint = "^2.10.2"
toml = "^0.10.2"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core==1.0.4"]

[tool.black]
exclude = '''
(
    /(
        \.git
        |\.tox
        |migrations
    )/
)
'''
include = '\.pyi?$'
line-length = 120

[tool.bandit]
targets = ["sources"]

[tool.isort]
profile = "deployer"
known_first_party = "deployer"

[tool.pylama]
format = "pycodestyle"
linters = "pycodestyle,pyflakes,mccabe,pylint"
max_line_length = 120
skip = "*/.tox/*,*/.env/*,*/migrations/*,.venv/*"

[tool.pylama.linter.pycodestyle]
ignore = "W503,W605"
max_line_length = 120

[tool.pylama.linter.pylint]
disable = "R,C,W,E1101"
max_line_length = 120
